hash benchmark

+++ src/base/types.hpp
@@ -2,7 +2,7 @@
 #include "config.h"
 #include "debug.cuh"
 
-typedef unsigned long long hash_t;
+typedef uint32_t hash_t;
 typedef unsigned long long index_t;
 typedef unsigned int index_s_t;
 typedef unsigned long long column_t;
